Add partial implementation of stub-inferring strategy
This strategy tries to infer type information from stub files.

The implementation is not yet fully functional.  See the
`test_stubstrategy.py` for cases that do not yet work correctly.
However, the implementation should not raise unwanted exceptions, but it
might not be able to infer all types from the stub files.  It does not
provide wrong ones, but only misses some cases.