Add Python 3.9 Docker image for CI
Experimentally add the latest Python 3.9 Docker image for unit test and
nightly test runs.  This currently uses the Python 3.9a1 version but
might automatically move to a later alpha or beta version as soon as
they are released and the according Docker file was adjusted on Docker
Hub.

We activate this as an experiment, and also to make sure that you
implementation won't break as soon as Python 3.9 will be released.
According to the release schedule [1], the beta phase of Python 3.9 is
planned to start on 2020–05–18 and the release candidate phase is
planned to start on 2020–08–10.  There is not yet a planned date for the
final release of Python 3.9.

This is truly an experiment, which might fail.  In case we suffer from
issues with Python 3.9, we might need to revert this commit and re-add
the changes later.