Fix module inspection and add test
The utility method `get_members_from_module` previously only returned
classes which was wrong since it should provide all members of a module.
It now provides all classes, functions, and methods, as detected by the
`inspect` module of the standard library.

Furthermore, we provide a test for this utility function that relies on
the `triangle` example.