Extract fixtures for module imports
Working with real modules is necessary for a couple of tests of the
generation algorithms.  Thus we create fixtures for the usage.  Both
fixtures are meant to be session-wide, i.e., their effect is only done
once.  This should not be a problem but more efficient, since it avoids
the reloading of modules for each and every test case.  It should not be
a problem as long as we only do read accesses on the fixture's return
values.

The fixtures purpose are the following:
- provide_imported_modules
  returns a dictionary of module names and the imported module content;
  the module name is only the last path of the fully-qualified name,
  thus be careful to not introduce duplicates when extending this
  fixture.
- provide_callables_from_fixtures_modules
  returns a dictionary of callable name to callable, which provides
  access to all classes, methods, and functions of the modules defined
  in the `provide_imported_modules` fixture.  Take care that there are
  no duplicate names for the callable names within the modules when you
  extend the modules!