Extract abstract executor
Execution cannot only be needed to check whether generated sequences are
executable, or which coverage they achieve, but also for other reasons.
An example would be to execute a type checker or a type inference
engine, such as mypy or MonkeyType.  The abstract class provides basic
functionality and an interface to allow such extensions of the
framework.