Randoopy: start algorithm incorporating MonkeyType
This algorithm is an extension of the Randoopy algorithm which utilises
the MonkeyType tool after a configurable number of iterations to execute
the successful test cases with MonkeyType's tracing activated.  This
shall be used to learn insights on potential parameter types of methods.
The implementation shall support to hand over the inferred information
to the existing `TestCluster`, such that the type information can then
be used in further iterations of the generation algorithm.