TypeInference: implement signature update
When new type information is inferred, the `InferredSignature` needs to
be updated to reflect these changes.  Previously, only the `parameters`
and `return_type` fields were updated.  This change also updates the
`signature` field according to the new type information.

Note that the `inspect.Signature` class is immutable according to its
documentation, thus an update of a signature will produce a new
`inspect.Signature` object every time.