Config: fix Algorithm enum declaration
The values attached to the enum constants need to be unique.  Since
every enum constants carries information about whether or not the
instrumentation shall be used, they cannot be unique if the user can
choose between more than two algorithms—which is the case.

By making the enum constant values being a pair of an auto-selected
integer and the flag whether or not instrumentation shall be used, they
are now unique again.