TestFactory: reuse existing variables for None types
A parameter with no type annotation can also get an existing value as
parameter.  Otherwise the generator only tries `None` as an argument
value, which is not helpful in most cases.  This is still far from being
perfect, because it relies on the fact that we were able to generate a
valid variable for at least one parameter, but it is a step in the right
direction.