RandooPy: do not generate Unions with NoneType
A union type with NoneType and another type is not very reasonable in
most cases.  Especially not, when it is because we learned a concrete
type for a variable, which was untyped before.  None might still be a
reasonable value and the generator should also try this value with a
small probability but it does not seem to be a reasonable type
annotation for most of the generation.