Add simple Docker file to generate container
The Docker container can be used to run `pynguin` as a standalone
application.  To use it, you need to mount the directories you need for
running to the container and use appropriate paths in `pynguin`'s
arguments.  All arguments that are given to `docker run` are directly
passed to `pynguin`.

An example of how to use the container can be like following:
```
docker build local/pynguin .
docker run \
    --mount type=bind,src=/path/to/source,target=/source \
    --mount type=bind,src=/path/to/result,target=/output \
    local/pynguin \
    --algorithm RANDOOPY \
    ...
```