Fix coverage result
By sending the current individual to the statistics tracker AFTER
storing coverage of the re-execution of the generated test suite, it
overrides the coverage value with the value of the last generated test
case.  This is obviously wrong, it must use the coverage from
re-executing the full generated test suite here.  Hence, we first send
the current individual and afterwards override the coverage with the
correct coverage value from the re-execution of the test suite.