Change final coverage calculation
To evaluate the final coverage values, we combine the passing and
failing test suite to one test suite and execute it.  The reason is that
we cover additional lines with the failing test suite, but throw these
test cases away, since they raise exceptions and we cannot decided
currently, whether these exceptions are on purpose or due to a
programming error.  The raising statements where covered, however, thus
should also be part of the coverage value.

Note that this change introduces a difference on purpose!  If you take
the passing test suite afterwards and execute it against the original
module under test, you might get a coverage value that is lower than the
final coverage value reported by Pynguin in the `Coverage' run-time
variable.  You'll get this value only by executing both the passing and
the failing test suite, though.