TestFactory: skip variational arguments
Changes the handling of how variational, i.e., `*args`, and keyworded,
i.e., `**kwargs`, parameters are generated.  Previously, we could not
handle them properly at all.  If we had a signature such as
`foo(*args, **kwargs)`, we generated two argument values for the two
arguments—which is conceptually wrong, since `*args` accepts a list of
arguments of undefined length and `**kwargs` accepts keyworded
arguments.

This change now skips those types explicitly, i.e., it does not generate
argument values for these.  It only logs their appearance and continues
with the next parameter.

This should be implemented in the future, together with the handling of
lists, sets, and dictionaries.