Executor: wrap execution for MonkeyType
In some rare cases, I was able to observe Exceptions raised from
MonkeyType when executing the statements, which was kind of surprising,
since these statements worked perfectly fine when executing them before
with our without coverage recording.  To avoid nasty failures, we simply
catch all raised exceptions and log them.

A possible consequence is that MonkeyType misses type information in the
current iteration.  Advantage is that the whole test-generation process
survives the exception and continues.