Cluster: fix wrong handling of inherited methods
Inherited methods should not be part of the test cluster.  Only methods
that are defined in a particular class should be, together with
overridden methods.

This change provides a regression test case exposing the issue, where
before it `__iter__` method was added to the methods of class `Bar`,
although only being inherited from `Foo`; it was furthermore not part of
the available methods of `Foo`.  The implementation now checks whether a
method was defined in the currently inspected class, if not it will not
be added to the cluster.

Furthermore, we provide a small readability refactoring for the skip
conditions.