Change post-dominator to dominator tree
Post-dominator and dominator tree are computed with the same algorithm,
the difference is only that for the post-dominator tree one has to
reverse the control-flow graph before computation.  This change adds the
possibility to compute both, with the underlying implementation being a
dominator tree.

This was suggested by a comment in merge request !15