CFG: filter for dead code nodes during creation
It appears that there are jump instructions in the byte code directly
after a return instruction, which cannot be called because the return
instruction already leaves the current method.  `bytecode`'s
`ControlFlowGraph` implementation, however, generates a node for these
dead-code nodes and tracks the jump target.  Our CFG generation then
takes this and adds a node to the CFG that has no predecessor but only
the jump target as a successor.

The underlying assumption and invariant on the CFG is that there is only
one entry node.  The above produces further entry nodes, which then let
the computation of, e.g., dominator trees fail.  We now assume all nodes
that have no predecessors and do not have the index 0, i.e., that are
not the original method entry, dead code and remove them from the graph.