CFG: add artificial entry node
There are situations where the CFG consists only of one block, which has
a self-loop, e.g., the following snippet:
```python
def receive(self):
    while True:
        print("True")
```
In such a case the implementation was not able to detect an entry node,
since there is no node that does not have any entering edges, i.e., our
invariant on an entry node.  We artificially insert one such node,
having index `-1` as a dummy for which the above-mentioned invariant
always holds.