Hacky fix for parameter expansion
When executing Pynguin in Docker on the cluster I had trouble with the
new space-separated parameters for the output_variables property.  This
is now a quick hack around to also use a comma-separated list for this
property, which will be expanded automatically when constructing the
Configuration object.

This is bad, I know, but it is a quick hack around the issue to be able
to run things on the cluster.