Draft a duck mock class
The duck mock utilises the idea of duck typing, i.e., a type A can be
used as a replacement for a type B if it provides the same methods, even
if they have no sub-typing relationship in the type hierarchy.

This is a first small version of the mock, it can be extended to provide
some kind of symbolic execution/restart for branchings; this would be
interesting for all comparison methods, such as `__eq__` or `__lt__`,
but also to return a new instance of this duck mock class as a return
value of methods to also track the information of those.