Apply fixit auto fixes
Instagram's new fixit linting framework [1] provides the option to not
only lint the code but also to apply the changes directly.  This commit
applies these suggestions.  We are currently not integrating fixit into
our build pipeline but better run it every now and then, because it is
very new and we suspect it to have a lot of changes in the next weeks
when it gets the attention of more and more developers.

[1] https://github.com/Instagram/Fixit