Chromosome: add checker for failing chromosome
A chromosome is a failing chromosome if its encapsulated test case is
failing.  This does obviously not hold for test-suite chromosomes, thus
they raise an exception here.

For future extensions were we want to consider the generation of
assertions for methods that raise an exception on purpose, we need to
refine this implementation.