Extract algorithm instantiation to factory
This introduces a factory to instantiate algorithms with all their
components.  The factory now needs to be filled with options (currently,
they are not available, but this can serve as a base for adding them).

The RandooPy strategy should be refactored anyway; it is now even more
hacky in its approach to extract the needed test-factory object from the
provided chromosome factory.  There are several more things in this
strategy that might need a future rework.