Enable usage of a configuration file
This change makes it possible to use a configuration file to set common
options and only change those options that are individual for a specific
tool run.

Note that this relies on `argparse`'s `fromfile_prefix_chars` feature,
i.e., one can give a file path prefixed by an `@` symbol that contains
options.  See the documentation of `argparse` for details.

Also note that this relies on the order in which the options are given.
That is, if the file is included last, it might override options set
before.  Therefore, I strongly recommend to put the configuration file
as the first argument and set all other arguments afterwards.  By doing
this, arguments on the command line will override arguments from the
configuration file.  This is, however, standard behaviour of `argparse`.