Docker: fix build and run in non-privileged user
The build stage of the Docker container actually did not build Pynguin
but relied on a build being available in the file system—which would the
user require to install `poetry` and all dependencies on their system
just to run Pynguin.  This patch now also builds Pynguin inside the
container.

Furthermore, since Pynguin executes arbitrary code and it previously was
running as a privileged user inside Docker, this could lead to security
issues.  We now run Pynguin as a non-privileged user `pynguin` inside
the Docker container to avoid any security issues related to privileged
users inside containers.