Include the tests into coverage computation
Coverage.py recommends to measure coverage also on the tests not only on
the implementation code [1].  The rationale is to see if parts of the
tests are not even executed—for example because they are marked as
skipped.  To detect this, measuring coverage of the executed test cases
is a simple, yet practical, solution.

Ideally, coverage on the test tree would be 100%.  In practice this does
not seem to be practical because some branches might not be coverable
for a good reason.

[1] https://twitter.com/coveragepy/status/1364642550573899781