Make pytest a main dependency
Generated tests might rely on pytest for the approximation of floats.
Therefore, pytest needs to be installed, to enable Pynguin to execute
the tests.  If a user installs Pynguin from PyPI or without the build
dependencies—which is perfectly fine, if they just want to use Pynguin
but not develop for it—they cannot execute tests otherwise.