Hotfix: manually close queue in executor
The queue is closed by the garbage collector but since we cannot
influence when the garbage collector runs this can cause trouble: each
open queue allocates resources, hence the number of open resource
handles can reach the OS's limit.  In this case, an OSError is thrown by
the implementation of the queue, which causes Pynguin to fail.

By explicitly closing the queue we can prevent this issue and free the
resources immediately after not needing them any more.