Add CITATION.cff file
The citation file format is an attempt to provide citation metadata in a
way that can be easily interpreted by humans and machines.  GitHub
supports this for a couple of months now—it adds a “Cite this
repository” link to the repository's side bar.

In this change, we add a basic file that names Pynguin's authors (as far
as their names are known), the DOI of the automatically pushed Zenodo
artifact, as well as a citation of our 2020 SSBSE paper on Pynguin.
Feel free to add further metadata in the future, e.g., further author
information and further references.

For details on the CFF format, see their GitHub page [1] and their
documentation [2].

[1] https://github.com/citation-file-format/citation-file-format
[2] https://citation-file-format.github.io/