Install `git' into Docker image
Pynguin needs a patched version of MutPy [1] to generate the assertions
based on mutation testing.  This fork is currently located in its own
GitHub repository [2], which is included as a dependency in the
`pyproject.toml` file.

In order to be able to install Pynguin, one needs to pull in the
dependencies, including the one from GitHub.  This can only be done,
when the `git' program is present in the image, thus we install it
explicitly.

[1] https://pypi.org/project/MutPy/
[2] https://github.com/f-str/mutpy