Guard imports for type checking
Many imports are only needed for type checking but not at runtime.  This
caused issues with circular imports and the like in the past.  To
resolve this, these changes guard all pynguin-internal imports that are
only necessary for type checking by `typing.TYPE_CHECKING` [1]; they
will only be imported for type checking but not for running Pynguin.

In order to make this work, the `from __future__ import annotations` is
necessary such that the type annotations are handled as strings and not
as objects [2] by static-analysis tools, such as pylint.  This will be
the default behaviour in future Python versions; it is not in Python 3.8
and 3.9, thus this additional import is necessary.  Alternatively, one
could have written the annotations as strings, which is more annoying in
my opinion.

[1] https://docs.python.org/3/library/typing.html#typing.TYPE_CHECKING
[2] https://www.python.org/dev/peps/pep-0563/