Fix crash of seeding for native modules
The constant seeding relies on the Python AST and requires that the
module names are Python files.  This is not the case for native modules,
such as compiled C extensions; those caused crashes of the seeding.  By
considering only Python files we fix these crashes.

Fixes https://github.com/se2p/pynguin/issues/20