Fix failing tracing of runtime type
For assertion generation, Pynguin tracks the variable types to decide
for which values it actually is able to generate assertions.  This
induced a bug that cause Pynguin to crash: creating an assertion on a
generator function, as given in the example in `is_ignorable_type`s
docstring does not work.  Hence, we omit this type for the set of
candidates we can generate assertions for.