Replace empty annotation and `object` by `Any`
For the test generator, all three are essentially the same: it can be
any type.  To avoid distinct handling of all three, we unify them in the
parsing step.  The idea is to simplify the handling code during
statement generation because we only need to deal with `Any`.  This
should also solve the problem having `None` vs `NoneType` as
type-values.