venv noob friendly changes
I'm newer to python and venv is a learning curve for me. Tried to elegantly refactor your docs to account for my learning curve, in the case others are in the same position. 
Reworded a few things that I had to read twice.
Created a Warning banner for the venv steps, and placed it before the pip command - as I ran the command, and then read the warning. Better to read warning before the command.