Remove splitting into passing and failing test suite
Splitting the generated test suite into a passing and a failing suite is
an artefact from the early days of Pynguin and evolves from the random
algorithm.  In the past, we had already changed the signature of the
generation algorithms and did this split after the generation.

With the improved output generation for exceptions we now get rid of
this and only export one test suite.  As a follow up, we shall
distinguish between expected and unexpected exceptions.