Make test cases distinguish between (un)expected exceptions
We consider an exception to be expected if it is explicitly raised in
the function's body or mentioned in its docstring.  For such exceptions,
Pynguin will generate a
```
with pytest.raises(Exception):
    ...
```
block in the test cases.

Every other exception that was raised during test-case execution is
considered an unexpected exception.  For those exceptions we will not
generate the aforementioned `with`-block but add the
`@pytest.mark.xfail` decorator to the test method to signal that this
test case is expected to fail.