Fix handling of nested functions in module analysis
Extracting the function description from the AST for a function with a
nested function (closure) causes the implementation to return two
function-description objects: one for the outer and one for the inner
function.  An assertion was checking that there is only one function
description as the result of this AST analysis.

This is correct in the sense that we are only interested in the analysis
of the outer function (since the inner cannot be called directly).
However, this requires additional filtering: we only keep the function
description of the outer function.  We can identify the outer function
by the name attribute of the analysed AST node (which is an (async)
function node)) and compare it to the name attribute of the
function-description object; we only keep the latter if the two names
are equivalent.