Exclude empty enum classes from test cluster
Parsing the included modules raised an issue when the `enum` module is
used: the test cluster then had a reference to the `enum.Enum` class,
which obviously does not contain any fields.  In the following,
generating tests failed, as soon as this class was selected to fulfil
parameter values because there was no field to select from, e.g.
`MyEnum.MY_FIELD`.

We now exclude empty enums from the test cluster to fix this issue.