Type inference selection compares the wrong objects
The object behind config.configuration.type_inference is an instance of
a data class, which itself hold the type-inference strategy in a field.
The comparison thus always fell through to the default case and thus
used the type-hints strategy as its default.  We now fix the comparison
and introduce a regression test to prevent this in the future.