Hotfix for type-handling issue
When deactivating type inference, we set all types to Any, which seems
to be a reasonable choice in the first place—since we do not know
anything about them, they could be anything.  This, however, turns out
to have some consequences on the statement factory, thus we revert this
change for now until we can properly fix this.