Omit unknown cyclomatic complexities
Our computation of the cyclomatic complexities requires the AST of a
function.  If it is not available, the complexity is yielded as -1 to
signal this.  However, when computing min, max, mean, and median
complexities, incorporating the -1 values does not make any sense—they
came from functions we were not able to analyse, e.g., an implicit
default constructor.  Hence, we omit those values.