Let `Coverage` contain the coverage value of the optimisation function
The recently introduced !175 (merged) (see also #200 (closed)) allows
to yield coverage values besides those that we used for the search,
e.g., we use branch coverage for optimisation but yield line and branch
coverage of the resulting test suite.

When implementing this it felt like a good choice to set the Coverage
output variable to a combination of all the yielded coverages.  However,
the more I think about it, it feels like not the best choice: during
optimisation we track a CoverageTimeline as well, which consists of the
coverage type the search optimises for; an obvious assumption is that
the last value of the timeline should be more or less equal to the value
of Coverage (at least very close, since the timeline is computed based
on some interpolation).

In my opinion it would be more intuitive if the Coverage variable
contains the final coverage value of the type that was used during the
search and the user can then select the LineCoverage or BranchCoverage
variables to retrieve line and branch coverage, respectively.