Fix a false-positive warning from PyLint
For some reason, at the call-site of this method (in method
update_return_type), the manual clearing of the cache leaded to a
warning from the latest PyLint version.  Setting the default argument
value here does not change any functionality, but just pacifies PyLint.
Seems to be more clear than a PyLint-disable comment.