Attempt to fix the read-the-docs build
Seems like RTD by default installs a very old version of their theme,
which depends on an old version of Jinja2 for the HTML template.
However, Pynguin relies on the latest version of Jinja2 for its HTML
coverage report.  This causes the installation of the latest Jinja2
version but an old RDT theme version when building the documentation on
RTD.  By now pinning the RTD theme version to its latest and setting it
as the default theme, we attempt to mitigate the issue and make the
build work again.