Fix area-under-curve computation
In case the maximum search time hits in the middle of a
generation-algorithm iteration, the last coverage value, i.e., the
coverage value of the population resulting from that iteration, will be
tracked _after_ the timeout.  Thus, dividing the area under curve by the
search time can lead to areas larger than 1.0 if being normalised.  By
not taking the search time into account for normalisation but the time
stamp of the last tracked generation, we shall mitigate this issue.