refactor: Wrapped importlib.reload in a try/catch temporarily
Problem: While testing tensorflow, reload triggered exceptions in the SUT because tf saves global states when importing, and reload makes the SUT save those global states again which tf does not allow and throws exceptions.
However, we don't know how not executing reload affects the rest of the code flow so this solution needs to be reevaluated.