Adjust default parameter values
We did a study on hyperparameter tuning [1].  In that study, we found
that for DynaMOSA some parameters did yield better performance than the
previously used default ones.  We now adjust the respective parameters
to the values that we found via tuning.

S. Lukasczyk and G. Fraser. 2025. “Search-based Hyperparameter Tuning
for Python Unit Test Generation”.  Symposium on Search-Based Software
Engineering (SSBSE).

[1] https://arxiv.org/pdf/2510.08716